bash: Copying into a new path on the Chromebook somehow changes the permission from 100755 to 100644. For these bash files that is fine as they aren't supposed to be executed by themselves.
bash: Copying into a new path on the Chromebook somehow changes the permission from 100755 to 100644. For these bash files, such a permission change might be fine as they aren't supposed to be executed by themselves.

bash: Copying into a new path on the Chromebook somehow changes the permission from 100755 to 100644. For these files, making them non-executable is not a problem as they aren't supposed to be input into another executables.